[BUGFIX] Cache various where clauses of PageRepository
Since TYPO3 v9, PageRepository works independently
from TSFE and can and should be instantiated on its
own, depending on the context.

This leads to multiple initializiation calls which can
and should be cached, as a page with 100 links will
re-trigger the properties all the time, even if they
are not needed (groupWhereClause for instance).

Caching these values reduces the time needed for
typolink generation a lot.

Resolves: #91208
Releases: master, 9.5
Change-Id: Ied7e2b78519754af2e10eeb4e9c1c9fcbbffe2e8
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/64321
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>