[TASK] Replace <style> for compliance with CSP header in svgs
SVG with inline styles are not loaded when the Content-Security-Policy
header contains the widely used setting "style-src 'self';" because
the browser has to reject interpreting the style attribute.
In this case a black rectangle is displayed.
Using attributes instead of styles is compliant with CSP "style-src
'self';" and the svgs can be displayed.

Releases: master, 10.4
Resolves: #91216
Change-Id: I1679a2004f1fc868231318249a267cbba62c540f
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/67693
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Ute Flierl <ute.flierl@helm9.de>
Tested-by: Richard Haeser <richard@richardhaeser.com>
Reviewed-by: Ute Flierl <ute.flierl@helm9.de>
Reviewed-by: Richard Haeser <richard@richardhaeser.com>