[BUGFIX] Show correct language title for inconsistent content
Show the correct language title in the new fluid page module if an
inconsistent content state has been detected. The title of the
problematic language must be shown instead of the default language
title.

Furthermore a not needed sprintf call is removed as the message itself
doesn't contain any placeholder which needs to be replaced.

Resolves: #91313
Releases: master
Change-Id: Ie1082bee58cb04068e2cea3b4f18c2b6f2b516f1
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/64416
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Daniel Goerz <daniel.goerz@posteo.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Josef Glatz <josefglatz@gmail.com>
Reviewed-by: Daniel Goerz <daniel.goerz@posteo.de>