[BUGFIX] Remove wrong MethodCallStaticMatchers
Some methods have been added for this matcher,
which are actually not deprecated/removed as a
whole. Only the usage of those methods has been
adjusted.

The extension scanner is not capable of detecting
such usages only, hence there is no sense in
reporting every usage of those functions, albeit
these usages might be valid.

The matcher entries are removed therefore.

Resolves: #91355
Releases: master
Change-Id: I9da87ecb320f65d4fe5df168d788bb2ba8547f84
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/64451
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Alexander Schnitzler <git@alexanderschnitzler.de>
Tested-by: Benjamin Franzke <bfr@qbus.de>
Reviewed-by: Alexander Schnitzler <git@alexanderschnitzler.de>
Reviewed-by: Benjamin Franzke <bfr@qbus.de>