[BUGFIX] Open CSH with selected context provided by links
A context can be provided, when opening the CSH (Context Sensitive Help).

E.g. when opening the CSH for a backend module or specific table field,
the help entry for that module or field will be opened.

This patch restores the described functionality by adding the action
to the link opened via JavaScript.

The "see also" links, used for cross referencing different CSH entries
are fixed as well. Cross referencing links are now build using the proper
ViewHelper to use backend module routing, instead of extbase routing.
This ensures arguments are not moved into an arbitrary extbase plugin
namespace.

Resolves: #91370
Releases: master
Change-Id: Ib6361e5a5f4ef441e098a595fa344f484a07ddc0
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/64477
Reviewed-by: Daniel Goerz <daniel.goerz@posteo.de>
Reviewed-by: Josef Glatz <josefglatz@gmail.com>
Reviewed-by: Sebastian Klein <laitnin@gmx.net>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Sebastian Klein <laitnin@gmx.net>
Tested-by: Josef Glatz <josefglatz@gmail.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>