[BUGFIX] Loosen site check for records not stored in site context
When saving a record on a page that is not part of a site,
the slug field of this record, despite being set to "uniqueInSite"
is not checked for uniqueness, as it is assumed unique enough.

This assumption needs to be applied as well when resolving
the record, instead of assuming the resolved record is not part
of the current site.

Releases: master, 9.5
Resolves: #91438
Change-Id: I347909b9b4caa523de3ad8e5d84c465e5d57b052
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/64525
Tested-by: Oliver Hader <oliver.hader@typo3.org>
Tested-by: TYPO3com <noreply@typo3.com>
Reviewed-by: Oliver Hader <oliver.hader@typo3.org>