[BUGFIX] Optimize database pre checks within install process
During the install process, the user has to choose a database in step 3
from a list of available databases.
If the database user has read access but no write access to the database,
the install tool used to throw an exception.
Now the system catches this exception, calculates permissions more accurately
and also provides more meaningful messages to the user.

Resolves: #78885
Releases: master, 10.4
Change-Id: Ib9aa632e8ca161acf2de628806386bd6607afd74
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/64564
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Richard Haeser <richard@richardhaeser.com>
Reviewed-by: Jörg Bösche <typo3@joergboesche.de>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>