[BUGFIX] Reintroduce tt_content sorting to both page modules
With #90235 db based sorting in the main query that fetches
the content elements to render has been accidentially removed.
In many cases, this did not have any impact: The sorting field
is part of a db index that is used for the query and some
dbms like maria/mysql then tend to return rows determined by
that index. But it can go wrong, too.
The fix is to add proper ordering again of course.

Resolves: #91223
Related: #90235
Releases: master, 10.4
Change-Id: If0931d8857a0e981181c72cc1a746d0a6b13ae12
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/65427
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>