[!!!][TASK] Remove CSS/JS inclusion via PHP in TSFE
TypoScriptFrontendController has had functionality to dynamically
load JavaScript and CSS code, also in non-cacheable (USER_INT) content.

This has been largely superseded via the AssetCollector API and
PageRenderer API which is built in a way to deal with non-cacheable
and cacheable content with proper method calls.

For this reason, the legacy functionality is removed:
* TSFE->setJS (used via TSFE->setJS('openpic') in core)
* TSFE->additionalJavaScript
* TSFE->additionalCSS
* TSFE->JSCode
* TSFE->inlineJS

Although all methods and properties have been marked as internal
in TYPO3 v10, a Breaking RST is still added.

Resolves: #91563
Releases: master
Change-Id: I167dae29e3daa5e1d5ad4c6848dcc49fbfec3183
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/64660
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Benni Mack <benni@typo3.org>