[BUGFIX] Use hashed object identifier in IRRE for nprogress
nprogress is used to indicate activity when a collapsed IRRE node is
about to get loaded. In case of FlexForms the id attribute may contain a
dot which is not an issue if handled correctly.
However, nprogress doesn't treat this value as an id, but rather as a
full CSS selector which causes issues and breaks loading the IRRE node.

To work around this issue, the id of the container used to render the
progress bar is now MD5 hashed.

Resolves: #91585
Releases: master, 10.4
Change-Id: I893d0cf24ea0f384d9ffff4d84f83b0fa35341b7
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/64783
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Daniel Goerz <daniel.goerz@posteo.de>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Oliver Hader <oliver.hader@typo3.org>
Reviewed-by: Daniel Goerz <daniel.goerz@posteo.de>