[BUGFIX] Register workspace preview middleware earlier
The current workspace preview middleware runs at a point
after the Page ID (and PageArguments) has been resolved by
the URL already.

This means a workspace preview with a preview user (and not
a backend user) does not work, as the user is registered too
late in the Middleware workflow.

In order to restore the functionality, the preview user is now
created BEFORE the PageResolver middleware, and does
everything (incl. sending the preview cookie, which is now
attached to the response, and not done during processing of the request).

However, as the webmount of the current page ID needs to be registered,
a second middleware is added which runs before TSFE->determineId()
but after the PageResolver and PageArgumentValidator middlewares.

Resolves: #91662
Releases: master, 10.4
Change-Id: Ic0108d2cd468f3ecf84e5a0e06c0fd5329046606
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/64900
Tested-by: Oliver Hader <oliver.hader@typo3.org>
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: David Steeb <david.steeb@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Oliver Hader <oliver.hader@typo3.org>
Reviewed-by: David Steeb <david.steeb@b13.com>
Reviewed-by: Benni Mack <benni@typo3.org>