[BUGFIX] Fetch TYPO3 version in reports module only
The core version check is run synchronously while logging
into the Backend. If the get.typo3.org service is down,
logins take extremely long.

Since there is no API to run this check asynchronously,
the patch drops this information from the toolbar,
it is only available in the reports module.

Resolves: #91507
Related: #84549
Releases: master, 10.4, 9.5
Change-Id: I78fd381df1aff6a40cc9bd57ac5a4babed2ec999
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/65160
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Benni Mack <benni@typo3.org>