[BUGFIX] Add max width for social media images
The og:image and twitter:image meta tags are not resized.
If an editor puts a 5000 pixel wide image in a field,
it does not get resized, and if it is large
enough (in size + MB) it doesn't get picked up everywhere.

For this reason, we'll limit the image width to 2000

Resolves: #91687
Releases: master, 10.4
Change-Id: Iec7b185c37b3685f523a0debebf90dbc7025bbc6
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/64909
Tested-by: Richard Haeser <richard@maxserv.com>
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: David Steeb <david.steeb@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Richard Haeser <richard@maxserv.com>
Reviewed-by: David Steeb <david.steeb@b13.com>
Reviewed-by: Benni Mack <benni@typo3.org>