[BUGFIX] Use URI of current request for referrer check
Instead of generating the URL to the backend login, the URI of the
current request is now used for the referrer check in backend login.
This fixes a redirect issue with password recovery links opened via
email.

The anchor-based reload detection has been replaced with a
localStorage-based solution as browsers don't trigger a new request if
the target location is already loaded, but only an achor is appended to
the URL.

Resolves: #91442
Releases: master, 10.4, 9.5
Change-Id: I577bdd8ce75c94f864852f812c0b8ad66f0d5634
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/65003
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Oliver Hader <oliver.hader@typo3.org>
Reviewed-by: Jörg Bösche <typo3@joergboesche.de>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>