[!!!][TASK] Store transOrigDiffSourceField as json string in db
The TCA ctrl transOrigDiffSourceField database field,
often called 'l18n_diffsource', stores information if
a default or source language record of a translated record
has been changed. This is used to hint editors with
"The record you translated from has been changed,
maybe an update of the translation is needed" in FormEngine.

The storage format of that field is a serialized PHP
string. This is no active, but a potential, general security
risk. To mitigate this, the storage format is changed to a
json string with the patch, including an install tool
database updater taking care of existing rows.

Resolves: #91906
Releases: master
Change-Id: Ib7e2a6128730eebd941ec8a565e72fd7a05d225d
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/65068
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>