[BUGFIX] Translation wizard shows correct elements in workspace
Since #89555 versioned records do not have pid=-1 set anymore,
but the uid of the live page or live placeholder record.
The translation modal of the page module to create records for
a translated page now shows records twice.
Solution is to use the correct database restriction to suppress
retrieval of versioned records of the translation source, those
are handled via BackendUtility::workspaceOL() already.

Resolves: #91851
Related: #89555
Releases: master, 10.4
Change-Id: Ia5262d3b25687c96f68389f5e8b49ec287c094e6
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/65029
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>