[BUGFIX] Provide example on how to limit phpunit test execution
The famous runTests.sh script misses an example on how to run
a filtered set of tests only in a given test file. This comes
in handy, when the wanted method name is not unique in the set
of tests.

By adding an example to the help text of the script, users
can run only their wanted test more easily.

Resolves: #91856
Releases: master, 10.4, 9.5
Change-Id: I2cf5c6f3e8d29ed9505bc52b10be3e22c65e0842
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/65080
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Simon Gilli <typo3@gilbertsoft.org>
Tested-by: Daniel Goerz <daniel.goerz@posteo.de>
Reviewed-by: Simon Gilli <typo3@gilbertsoft.org>
Reviewed-by: Daniel Goerz <daniel.goerz@posteo.de>