[TASK] Allow typo3/cms to be installed with Composer 2
Currently the dev dependency ocramius/package-versions does not
work with Composer 2. To overcome this, Composer published
an API compatible package which can be used as replacement.

This does not affect any published package. It is only
a preparation to be able to switch to Composer 2
for testing and packaging.

composer req --dev composer/package-versions-deprecated --no-update
composer up typo3/cms-composer-installers composer/package-versions-deprecated ocramius/package-versions

Releases: 10.4, master
Resolves: #91864
Change-Id: I943cb07486f444c1d971afcde37c0d5fe2becd13
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/65085
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Simon Gilli <typo3@gilbertsoft.org>
Tested-by: Frank Nägler <frank.naegler@typo3.org>
Reviewed-by: Manuel Selbach <manuel_selbach@yahoo.de>
Reviewed-by: Simon Gilli <typo3@gilbertsoft.org>
Reviewed-by: Frank Nägler <frank.naegler@typo3.org>