[BUGFIX] Reset Type Conversion Messages after usage
PropertyMapper (which is a singleton) previously created a new
$messages instance every time "convert" was called.

This works for most cases, however when using "getMessages()"
only the last message was returned.

It would be very useful to keep the messages in place
until "resetMessages()" is called, which is used within each
action of the Extbase controller validation
to avoid some edge-case issues when using multiple actions on one request.

Resolves: #70872
Releases: master
Change-Id: I2fddc4261567cad2ebb12f91296ff3ec8d681ad8
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/65114
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>