[BUGFIX] Rename default storage from "fileadmin/ (auto-created)" to "fileadmin"
The information in the sys_file_storage record created automatically
should not be persisted as "auto-created" as 90% of the people do not
modify this, and it is only confusing to see this information in the filelist
without having a good and explained way to change this.

Instead "fileadmin" (taken from $GLOBALS['TYPO3_CONF_VARS']['BE']['fileadminDir'])
is used when creating the default storage.

This was originally added by myself when introducing FAL in 6.0,
and never touched, however feedback from newcomers shows that
this should be adapted.

Resolves: #91921
Releases: master
Change-Id: I675ef7818aa6a83191574131f54c1c2b6f490e41
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/65151
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>