[BUGFIX] Workspace module: Clear row selection after publish
The row selection of records in JavaScript is not reset after
execution of a publish action in the workspace module.
If an editor tries to publish or disard another set of
records without reloading the module, the already handled
records from the first action are send again in the request
payload, leading to server side exceptions.

Resolves: #89927
Releases: master, 10.4
Change-Id: I8800c1df34fd749d84e63ea3e42c7b83513e1a38
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/65170
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>