[BUGFIX] Always use 'paste' wording pasting clipboard elements
The clipboard has a 'move' and a 'copy' mode. This can be
changed in the list module where the clipboard is rendered.
Using 'cut' in context menu, the state is forced to 'move',
using 'copy' in context menu, the state is forced to 'copy'.

If there is an element in the clipboard, 'paste' icons in
list and page module appear. Their title tag is always
'paste' or 'paste after' or similar.

Only the modal in page module distinguishes between copy
and paste and changes its wording 'Do you really want
to move' vs. 'Do you really want to copy', while the
modal title is always 'paste'.

This is odd from an UX point of view: Users learned they
either 'cut + paste' or 'copy + paste': The decision to
cut or paste is the first step when selecting the element,
it's then either cut-out and moved to clipboard, or a copy
in the clipboard is added. The second step is then just
'paste here'.

The patch changes the modal wording to always use 'paste'
as second step, just like the icons already do.

Resolves: #83566
Releases: master, 10.4
Change-Id: Ia7b4296bbcee1d21b2be0d0ddb9e1019c93e9b4a
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/65199
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>