[BUGFIX] Output messages and progress for language:update
Always output a progress bar and information that it is
actually doing something even when not called with --verbose

When no output is required or wanted, the global option --quiet
can be used instead.

To further improve the command, --no-progress is added
which removes the progress bar output, and shows messages instead.

Last but not least, output success or error messages for each
of the language pack fetches and exit with error code when
failures occurred.

To not make the command always fail when private extensions are used,
individual extensions can be skipped.

Releases: master, 10.4
Resolves: #91988
Change-Id: Ie7b7717f7f6fc2f884e974dbca564ca5a2c013af
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/65289
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Jörg Bösche <typo3@joergboesche.de>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Jörg Bösche <typo3@joergboesche.de>
Reviewed-by: Benni Mack <benni@typo3.org>